Funca solo falta que borre los comentarios

+++ Application/Interfaces/Infraestructure/Query/IAcompananteQuery.cs
@@ -4,7 +4,7 @@ namespace Application.Interfaces.Infraestructure.Query
 {
     public interface IAcompananteQuery
     {
-        Task<List<Acompanante>> GetAcompananteFiltros(int? Id, int? Especialidad, int? Disponibilidad, int? ObraSocial, string? ZonaLaboral);
+        Task<List<Acompanante>> GetAcompananteFiltros(int? Id, int? Especialidad, Int16? Disponibilidad, int? ObraSocial, string? ZonaLaboral);
         Task<List<Acompanante>> GetAcompanantes();
         Task<Acompanante?> GetAcompananteById(int Id);
     }
